Move _config.yml into docs/
The _config.yml file doesn't seem to be *doing* anything.
I think the problem is that it's in the root directory,
not in the source root, so let's move it to the source root
and hopefully that fixes the problem. I don't really have a way
to independently test this, so let's just do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>