Tweak guide now that it's posted to the website
Make a number of changes to the guide, including explaining
how to download the latest version (the user may have printed it)
and putting the screenshot into the guide itself.

I changed the heading levels of all the reference material
so they're all under a "reference" section, that makes it
easier to understand the overall document structure.

I made copies of the license files so there are 2 copies.
That way, they're in the "expected" place (the root dir)
and a directory that's going to the website (the source dir).
Having 2 copies isn't ideal, but we don't expect them to change
and this way we are certain to have a copy of the license handy.
I doubt anyone will contest the license, but I always want to
make licensing crystal-clear.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>