Reorder reference manual "editor tab" section. Fixes #63.
This reorders the reference manual "editor tab" section.
I tried to create a "logical" approach, but that ends up
constantly making forward references.

This version starts with a simple top-to-bottom view of
the major parts of the editor tab UI, then explains
hypotheses and goals, then goes into detail on some dialogues.
This eliminates the forward references and is easier to
understand as a reference manual section.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>