Use dagger instead of statement/step notation
It's confusing to say "statement" when you mean "step", but
it's *also* confusing when you claim the UI says one thing but the
real UI says something different.

I tried to solve this with statement/step but that was too confusion.
Let's try "statement‡" instead to indicate "we're using statement
to be consistent with the UI but we really mean step".

I made this change primarily with:

~~~~
sed -E -e 's/statement\/step/statement‡/g' \
       -e 's/statement\(s\)\/step\(s\)/statements‡/g'
       -e 's/statements\/steps/statements‡/g' < docs/index.md > docs/,1
~~~~

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>