Improve placement of double-daggers
Remove the double-daggers where we really do mean statements
not steps (e.g., in some areas of the bottom-up prover).

Add double-daggers to the use of "substitution"; my hope is that
the tool will show "replacement" everywhere eventually.
That way we can use the word "substitution" for the underlying
Metamath substitution operation, and "repacement" for the
editor menu item. It's always nice to consistently use two different
words for two different things :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>