Move "Export to URL" to later.
I expect TEMP mode is likely to have changes,
e.g., it may have a different visual display, and on loading
a URL the tool may allow you to choose to use TEMP mode or not
(instead of *always* using TEMP mode).

I plan to use the 2p2e4 section for a video, and I'd like to
focus on areas of the tool that are less likely to change
(or at least are likely to change less).

So let's move the discussion about "export to URL" and TEMP mode
to a later part of the document. It's fine if we discuss it later,
If a user accidentally ends up in TEMP mode, the tool will warn
the user if they try to do edits (at least direct statement edits),
and I've documented TEMP mode in the tutorial *and* reference sections.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>