When clicking on application page, open a *new* page
We want users to click on the application page, but by default
that means they'll leave the guide, which probably isn't what
they wanted to do while reading a guide.

So, modify the link so it switches to a new tab.
This also adds the rel=".." material that shouldn't be needed
by modern browsers as a security measure, but it doesn't hurt and
might help someone with an older browser.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>